Backport security and accessibility fixes
- disable Express x-powered-by via shared app setup\n- add release group button aria-label\n- hide decorative sidebar icons and badges from assistive tech\n- adapt security test to exercise shared app setup\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
